+++ python/marvin/web/lib/src/table.js
@@ -2,7 +2,7 @@
 * @Author: Brian Cherinka
 * @Date:   2016-04-25 13:56:19
 * @Last Modified by:   Brian Cherinka
-* @Last Modified time: 2017-05-24 10:48:06
+* @Last Modified time: 2017-06-04 02:03:56
 */
 
 //jshint esversion: 6
@@ -42,6 +42,7 @@ class Table {
         this.table.bootstrapTable({
             classes: 'table table-bordered table-condensed table-hover',
             toggle : 'table',
+            toolbar: '#toolbar',
             pagination : true,
             pageSize: 10,
             pageList : '[10, 20, 50]',